% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStatistics.R
\name{DescriptiveStatistics}
\alias{DescriptiveStatistics}
\title{DescriptiveStatistics}
\usage{
DescriptiveStatistics(d)
}
\arguments{
\item{d}{dataset of water quality in long format}
}
\value{
dataset showing descriptive statistics per parameter:
- number of measurements per parameter
- detection limit of parameter (if applicable)
- percentage of measurements per parameter that is below the detection limit
- mean, median, min, max and stdev of values per parameter
- skewness of measurements per parameter
}
\description{
Function to return descriptive statistics of hydrochemical data per parameter (Cl, Na, etc)
}
