% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Testdata hydrochemistry documentation}
\format{
A dataframe containing the relevant columns and data to use the functions in this package. Columns include:
- wellcode
- samplecode
- year
- parameter
- value
- sd
- limit_symbol
- detection_limit
- units
- method
- notes
- watercode
- sampletype
- subtype
- xcoord
- ycoord
}
\usage{
example_data
}
\description{
An example dataset containing a part of the hydrochemical data of Curacao to show how functions work.
}
\keyword{datasets}
