% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meql.R
\name{meql}
\alias{meql}
\title{meql}
\usage{
meql(d, corr = T)
}
\arguments{
\item{d}{dataset of water quality in long format}

\item{corr}{correction option for converting alkalinity as CaCO3 to HCO3}
}
\value{
dataset of concentrations in meq/L in wide format with Electrical Balance check
}
\description{
Function to convert hydrochemical dataset with concentrations in wide format to meq/L
and calculate Electrical Balance. Concentrations need to be in mg/L.
}
