% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PiperPlot.R
\name{PiperPlot}
\alias{PiperPlot}
\title{PiperPlot}
\usage{
PiperPlot(d, NO3 = FALSE)
}
\arguments{
\item{data}{dataset of water quality in long format}
}
\value{
piper diagram ggplot object
}
\description{
Function to produce piper diagram that displays water types of hydrochemical data.
Input concentrations need to be in mg/L.
}
\details{
Piper diagram functions adjusted from https://rstudio-pubs-static.s3.amazonaws.com/542159_ae160ba405044883a58ba3a53e4f7e6d.html
to include option for adding NO3 and water type legend to the ggpiper_plot object.

Not fully functional and thoroughly tested!!
}
